towards signed return values

+++ ringbuffer.h
@@ -69,13 +69,13 @@ void ringbuffer_init(ringbuffer_t* rb, uint8_t* mem, size_t memlen);
 /*
  * TODO: Add description
  */
-size_t ringbuffer_get_length(ringbuffer_t* rb);
+int ringbuffer_get_length(ringbuffer_t* rb);
 
 
 /*
  * TODO: Add description
  */
-size_t ringbuffer_get_space(ringbuffer_t* rb);
+int ringbuffer_get_space(ringbuffer_t* rb);
 
 
 /*
@@ -87,38 +87,38 @@ void ringbuffer_clear(ringbuffer_t* rb);
 /*
  * TODO: Add description
  */
-size_t ringbuffer_write(ringbuffer_t* rb, uint8_t* data, size_t len);
+int ringbuffer_write(ringbuffer_t* rb, uint8_t* data, size_t len);
 
 
 /*
  * TODO: Add description
  */
-size_t ringbuffer_read(ringbuffer_t* rb, uint8_t* data, size_t len);
+int ringbuffer_read(ringbuffer_t* rb, uint8_t* data, size_t len);
 
 
 /*
  * TODO: Add description
  */
-size_t ringbuffer_read_memory(ringbuffer_t* rb, uint8_t** data, size_t len);
+int ringbuffer_read_memory(ringbuffer_t* rb, uint8_t** data, size_t len);
 
 
 /*
  * TODO: Add description
  */
-size_t ringbuffer_peek(ringbuffer_t* rb, uint8_t* data, size_t len);
+int ringbuffer_peek(ringbuffer_t* rb, uint8_t* data, size_t len);
 
 
 /*
  * TODO: Add description
  */
-size_t ringbuffer_peek_offset(
+int ringbuffer_peek_offset(
         ringbuffer_t* rb, size_t offset, uint8_t* data, size_t len);
 
 
 /*
  * TODO: Add description
  */
-size_t ringbuffer_discard(ringbuffer_t* rb, size_t len);
+int ringbuffer_discard(ringbuffer_t* rb, size_t len);
 
 
 /* ========================================================================= */
